requirements.txt
numpy==1.23.5
pygame==2.6.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
